requirements.txt
Flask==2.3.2
mysql-connector-python==8.0.33
prometheus-client==0.20.0